Remove NormalTitleView visibility toggle in search methods
The code changes involve modifying the `ShowSearchView_Clicked` and
`HideSearchView_Clicked` methods in the `HomePageM.xaml.cs` file.
Specifically, the lines that set the visibility of `NormalTitleView`
have been removed. This means that when the search view is shown or
hidden, the visibility of `NormalTitleView` will no longer be toggled.
- Refactored LoadSongs
- Fixed #101