Update version, add null checks, refactor ShareSongPage
Updated the version number in the application title from "Dimmer v0.1.2" to "Dimmer v0.2.1" for both debug and release builds.
Changed the active debug framework and profile in `Dimmer-MAUI.csproj.user` from Android to Windows.
Removed the `PlaylistMenuBtmSheet.xaml` from the project file.
Added null checks for `TemporarilyPickedSong`, `SelectedSongToOpenBtmSheet`, and `song` in various methods to prevent null reference exceptions.
Refactored `ShareSongPage` to use a property named `HomePageVM` instead of `vm` and added null checks for `TemporarilyPickedSong`.
Modified the file save path format in `ShareSongPage` to remove seconds from the timestamp and added a check to delete the file if it already exists.
Ensured that the background color is set to transparent when removing a custom image in `ShareSongPage`.
Updated the method to load lyrics in `ShareSongPage` to use the new `HomePageVM` property.