Increase HeightRequest and update DXButton attributes
In `AlbumsM.xaml`, increased `VerticalStackLayout`'s `HeightRequest` from 290 to 380. Added a new line to the `dx:DXButton` to separate the `PressedBackgroundColor` attribute for better readability.

In `HomePageM.xaml`, increased `VerticalStackLayout`'s `HeightRequest` from 300 to 380. Updated the `dx:DXButton` within the `dx:BottomSheet` to include `PressedBackgroundColor` set to "DarkSlateBlue" and `ButtonType` set to "Filled".