Refactor SingleSongPage layout and data binding
Updated `SingleSongPage.xaml` to introduce a new layout using `Grid` and `SfTabView` for better organization of song details, including tabs for lyrics and song attributes. Removed the old layout and added a close button for the sidebar.

Modified the constructor in `SingleSongPage.xaml.cs` to accept a `BaseViewModelWin` parameter, setting it as the `BindingContext` to enable data binding and load song information from LastFM when the page appears.