Refactor SongModel and SongModelView properties
Removed unused byte array properties (`CoverImageBytes`, `ArtistImageBytes`, `AlbumImageBytes`) from `SongModel` and `SongModelView` to streamline the data model.

Added a new `UnSyncLyrics` property to `SongModel` to support additional lyrics functionality. Updated `SyncLyrics` in both classes to have a default value of `string.Empty` and marked it as observable in `SongModelView`.

Refactored `HasSyncedLyrics` in `SongModelView` to depend on the updated `SyncLyrics` property.