Update UI bindings and add FolderScanCompleted handling
Updated `MediaControlsView` visibility in `HomePage.xaml` to use a binding with a converter based on `SearchResults.Count`. Removed hardcoded `IsVisible="True"`.

Added a `TextBlock` to `SettingsPage.xaml` to display the latest app log (`LatestAppLog.Log`) at the bottom of the grid.

Enhanced `BaseViewModel.cs` by adding a subscription to handle the `FolderScanCompleted` state. Observes this state on the UI thread and logs any exceptions during the process.